fastapi_app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-core
langchain-anthropic
langgraph
